requirements.txt
fastapi
uvicorn
requests
groq
yfinance
reportlab
python-dotenv